Ensure dll gets updated version before build
