+++ .ado/bumpFileVersions.js
@@ -1,0 +1,8 @@
+// @ts-check
+// Just bump the local version numbers in the files...
+
+const {
+  updateVersionsInFiles,
+} = require("./versionUtils");
+
+updateVersionsInFiles();
